Support 278 Transaction files
Added some tests and small changes to ensure that transactions in
nominee 278s as well as 278 transactions (which have an extra column)
get extracted.

This isn't perfect as some of the PFDs are weird and the second part of
the extra long header on the 278T gets pulled in as the first row.